+++ color_tools.h
@@ -1,28 +1,28 @@
 
-void mult_matrices(double a[][3], double b[3], double result[3]);
-void print_matrix(double a[3]);
+void mult_matrices(float a[][3], float b[3], float result[3]);
+void print_matrix(float a[3]);
 void CIEmatricesLandR();
-void element_wise_multiply(double matrix[][3], double vector[3], double result[][3]);
-void CIELab_A_matrixes(double Al[][3], double Ar[][3], double CIEmatrix[][3]);
-double cielabf(double U, double Un);
-double Lstar(double x, double y, double z); 
-double Astar(double x, double y, double z);
-double Bstar(double x, double y, double z); 
+void element_wise_multiply(float matrix[][3], float vector[3], float result[][3]);
+void CIELab_A_matrixes(float Al[][3], float Ar[][3], float CIEmatrix[][3]);
+float cielabf(float U, float Un);
+float Lstar(float x, float y, float z); 
+float Astar(float x, float y, float z);
+float Bstar(float x, float y, float z); 
 
-double LCD_Lstar(double rgb[3]);
-double LCD_Astar(double rgb[3]);
-double LCD_Bstar(double rgb[3]);
+float LCD_Lstar(float rgb[3]);
+float LCD_Astar(float rgb[3]);
+float LCD_Bstar(float rgb[3]);
 
-double CLstarRight(double rgb[3]);
-double CAstarRight(double rgb[3]);
-double CBstarRight(double rgb[3]);
-double CLstarLeft(double rgb[3]);
-double CAstarLeft(double rgb[3]);
-double CBstarLeft(double rgb[3]);
+float CLstarRight(float rgb[3]);
+float CAstarRight(float rgb[3]);
+float CBstarRight(float rgb[3]);
+float CLstarLeft(float rgb[3]);
+float CAstarLeft(float rgb[3]);
+float CBstarLeft(float rgb[3]);
 
-double Lstarright(double rgb[3]);
-double Astarright(double rgb[3]);
-double Bstarright(double rgb[3]);
-double Lstarleft(double rgb[3]);
-double Astarleft(double rgb[3]);
-double Bstarleft(double rgb[3]);
+float Lstarright(float rgb[3]);
+float Astarright(float rgb[3]);
+float Bstarright(float rgb[3]);
+float Lstarleft(float rgb[3]);
+float Astarleft(float rgb[3]);
+float Bstarleft(float rgb[3]);
