+++ methods/modified_photoshop.c
@@ -14,9 +14,9 @@
 
 CvScalar method_combine_pixels(CvScalar left_pixel, CvScalar right_pixel)
 {
-    double left_greyscale = 0.114*left_pixel.val[0] +
-                                   0.587*left_pixel.val[1] +
-                                   0.299*left_pixel.val[2];
+    float left_greyscale = 0.114*left_pixel.val[0] +
+                           0.587*left_pixel.val[1] +
+                           0.299*left_pixel.val[2];
     //printf("%f %f %f %f\n", left_pixel.val[0], left_pixel.val[1], left_pixel.val[2], left_greyscale);
     return cvScalar(
         left_greyscale,
